requirements.txt
aiohttp>=3.12.9
discord.py>=2.6.3
google_api_python_client>=2.181.0
google_auth_oauthlib>=1.2.2
protobuf>=6.32.0
